+++ tests/ClientTest.php
@@ -3,11 +3,11 @@
 namespace NotificationChannels\Bandwidth\Test;
 
 use GuzzleHttp\Handler\MockHandler;
-use PHPUnit\Framework\TestCase;
 use GuzzleHttp\HandlerStack;
 use GuzzleHttp\Psr7\Response;
 use NotificationChannels\Bandwidth\BandwidthClient;
 use NotificationChannels\Bandwidth\BandwidthConfig;
+use PHPUnit\Framework\TestCase;
 
 class ClientTest extends TestCase
 {