+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+### 1.6.0
+* Add: throw custom exception on failure
+* Add: dispatch `NotificationFailed` event before throwing exception
+* Add: log Notification id in simulation
+
 ### 1.5.0
 * Add `simulate` option